id_api/pyproject.toml
[tool.poetry]
name = "id-api"
version = "0.1.0"
description = ""
authors = ["Abdelfattah Hamdi <abdelfattah.hamdy234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
djangorestframework = "^3.16.1"
redis = "^6.4.0"
django-redis = "^6.0.0"
drf-yasg = "^1.21.11"
celery = "^5.5.3"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.11"
psycopg2 = "^2.9.11"
dotenv = "^0.9.9"
python-dotenv = "^1.1.1"
whitenoise = "^6.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
